requirements.txt
certifi==2017.4.17
chardet==3.0.4
dateutils==0.6.6
dj-static==0.0.6
Django==1.11.2
django-ckeditor==5.3.0
django-crispy-forms==1.6.1
django-js-asset==0.1.1
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
googlemaps==2.5.0
html2text==2016.9.19
idna==2.5
Markdown==2.6.9
olefile==0.44
paho-mqtt==1.2.3
Pillow==4.2.1
pypiwin32==219
python-dateutil==2.6.0
pytz==2017.2
requests==2.17.3
six==1.10.0
speech==0.5.2
static3==0.7.0
urllib3==1.21.1
gunicorn==19.6.0
psycopg2==2.6.2